Pre-select picker value on both iOS and Android
It is reported that on Android, if the user open the picker and
select nothing and press done. The underlying callback is
actionSelect with selectedItem being undefined. This pattern
does not match nothing thus no callback is invoked. The state
comes to a dead end and the picker cannot be re-opened anymore.

The proposed fix is make the picker behaves in the same way as
the iOS one. We are going to pre-select.